+++ 이상민/SWEA/D3/20551. 증가하는 사탕 수열/README.md
