+++ 이상민/백준/Silver/11057. 오르막 수/오르막 수.py
@@ -1,0 +1,8 @@
+n = int(input())
+arr = [[1] * 10 for _ in range(n+1)]
+
+for i in range(1, n+1):
+    for j in range(1, 10):
+        arr[i][j] = (arr[i][j-1] + arr[i-1][j]) % 10007
+
+print(arr[n][9])
